python start
Former-commit-id: 6af73a46967280c9ca34124eea6d71a33ef3143b

+++ src/term.c
@@ -195,7 +195,7 @@ static void term_core_putc (const char c)
 
 static void term_core_puts (const char *str)
 {
-    const size_t l = strlen(str);
+    const int l = strlen(str);
 
     if (term_core_buffer_pos + l >= term_core_buffer_size) {
         while (term_core_buffer_pos + l >= term_core_buffer_size) {
